backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.10.0
groq==0.13.1
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.6
